samples: gst-inspect implementation
This commit introduces a first implementation of the GStreamer tool
'gst-inspect-1.0'. More precisely, it prints the result of the command
'gst-inspect-1.0 -a'. This is intended to be a temporal solution. In the near
future we should consider doing something with those ugly include directories,
like converting GStreamer to a subproject instead of a dependency.